+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - An advanced widget named `flowbox`.
 - `focusable` property to all widget.
 - `widget_name` property to all widget.
+- `lifetime` flag for update command.
 
 ### Fixed